chore(tsc): split `types/check.ts` into smaller files
