+++ types/response.ts
@@ -1,0 +1,65 @@
+import { reply } from 'worktop';
+import { ServerResponse } from 'worktop/response';
+import type { FetchHandler } from 'worktop';
+
+assert<FetchHandler>(
+	reply(event => {
+		assert<FetchEvent>(event);
+		assert<Request>(event.request);
+		return fetch(event.request);
+	})
+);
+
+// @ts-expect-error
+new ServerResponse();
+// @ts-expect-error
+ServerResponse('GET');
+
+const response = new ServerResponse('GET');
+
+assert<ServerResponse>(response);
+
+assert<Headers>(response.headers);
+assert<number>(response.statusCode);
+assert<BodyInit | null>(response.body);
+assert<boolean>(response.finished);
+
+// @ts-expect-error
+response.finished = true;
+
+assert<object>(response.getHeaders());
+assert<string[]>(response.getHeaderNames());
+assert<boolean>(response.hasHeader('foo'));
+assert<string | null>(response.getHeader('foo'));
+assert<void>(response.setHeader('foo', 'bar'));
+assert<void>(response.removeHeader('foo'));
+
+// NOTE: native cast to string
+response.setHeader('foo', 123);
+response.setHeader('foo', [123]);
+response.setHeader('foo', ['bar']);
+response.setHeader('foo', 123.45678);
+response.setHeader('foo', ['a', 1, 'b']);
+
+// @ts-expect-error
+response.setHeader('foo', { foo: 123 });
+// @ts-expect-error - altho technically ok
+response.setHeader('foo', new Date);
+
+// @ts-expect-error
+assert<void>(response.writeHead('foo'));
+assert<void>(response.writeHead(200, { foo: 'bar'}));
+assert<void>(response.writeHead(200));
+
+// @ts-expect-error
+assert<void>(response.end({ foo: 123 }));
+assert<void>(response.end(new FormData));
+assert<void>(response.end('123'));
+assert<void>(response.end(''));
+
+// @ts-expect-error
+assert<void>(response.send('200'));
+assert<void>(response.send(200, undefined));
+assert<void>(response.send(200, null, { foo: 'bar' }));
+assert<void>(response.send(200, new FormData));
+assert<void>(response.send(200));
